1. 删除实体、字段增加二次确认； 2. 表单设计页面允许创建（主实体）字段（暂时不支持创建明细实体字段）。

+++ src/views/system/entity-field-table.vue
@@ -408,7 +408,20 @@ export default {
 						showCancelButton: true,
 						type: 'warning'
 					}).then(() => {
-						this.deleteField(row.name, this.entity)
+						//再次确认删除
+						this.$prompt('请输入大写"MLC"', '再次确认删除', {
+							confirmButtonText: '确定',
+							cancelButtonText: '取消',
+							inputPattern: /^MLC$/,
+							inputErrorMessage: '输入不正确'
+						}).then(({value}) => {
+							if (value === 'MLC') {
+								this.deleteField(row.name, this.entity)
+							}
+						}).catch((err) => {
+							console.error('eeee', err)
+							this.$message({type: 'info', message: '已取消删除'});
+						})
 					}).catch(() => {
 						this.$message({type: 'info', message: '已取消删除'});
 					});
